requirements.txt
mistralai==0.1.8
pycodestyle==2.11.1
python-dotenv==1.0.1
neo4j==5.18.0
ruff==0.3.5